feat: support auto save

+++ pages/note.tsx
@@ -10,7 +10,7 @@ import TreeItem, { treeItemClasses } from '@mui/lab/TreeItem';
 import CodeEditor from '../components/Note/Editor';
 import MarkdownView from '../components/Note/Preview';
 import { TransitionProps } from '@mui/material/transitions';
-import {useSpring, animated} from 'react-spring'
+import { useSpring, animated } from 'react-spring'
 import { createDoc, deleteDocs, readAllDocs, saveDoc } from '../utils/api';
 import { NoteNode } from '../types/type';
 import useLoading from '../hooks/useLoading';
@@ -18,11 +18,12 @@ import { LoadingButton } from '@mui/lab';
 import MoreHorizIcon from '@mui/icons-material/MoreHoriz';
 import DropDown from '../components/DropDown';
 import { useLocalStorage } from 'react-use';
+import { toast } from '../components/Toast';
 
 enableMapSet();
 
 const getName = (str: string) => {
-  const res = str.match(/#\s+(.+)(?![\r\n]+)/);
+  const res = str.match(/#\s+(.+)[\r\n]+/);
   if (!res) return;
   return res[1];
 }
@@ -77,7 +78,7 @@ const searchState = atom<string>({
 
 const currentNoteSelector = selector({
   key: 'currentNoteSelector',
-  get: ({get}) => {
+  get: ({ get }) => {
     const noteNodesMap = get(noteNodesMapState);
     const currentNoteId = get(currentNoteIdState);
     if (!currentNoteId) return null;
@@ -93,10 +94,10 @@ const noteNodesMapState = atom<NoteNodesMap>({
 
 const notesTreeSelector = selector({
   key: 'notesTree',
-  get: ({get}) => {
+  get: ({ get }) => {
     const noteNodes = get(noteNodesMapState);
     const roots = [];
-  
+
     for (const note of noteNodes.values()) {
       if (note.pid === null) {
         roots.push(note);
@@ -119,11 +120,41 @@ const notesTreeSelector = selector({
   }
 })
 
+const needSaveState = atom({
+  key: 'needSaveSelector',
+  default: false,
+})
+
+const AUTO_SAVE_TIME = 13;
+
+const autoSaveCountDownState = atom({
+  key: 'autoSaveCountDownState',
+  default: AUTO_SAVE_TIME,
+})
+
+const autoSaveTimerState = atom<number | null>({
+  key: 'autoSaveTimerState',
+  default: null,
+})
+
+const useCheckNeedSave = () => {
+  const needSave = useRecoilValue(needSaveState);
+  return () => {
+    if (needSave) {
+      toast('save current editor!', { alertProps: { severity: 'error' } })
+      return true;
+    }
+    return false;
+  }
+}
+
 const ActionBar: React.FC = () => {
   const [loading, withLoading] = useLoading()
   const setNotesState = useSetRecoilState(noteNodesMapState)
   const [searchValue, setSearchValue] = useRecoilState(searchState)
+  const checkNeedSave = useCheckNeedSave();
   const handleAdd = async () => {
+    if (checkNeedSave()) return;
     const res = await withLoading(createDoc(id => createNoteNode(id)))
     setNotesState((origin) => produce(origin, draft => {
       draft.set(res.id, res)
@@ -182,24 +213,30 @@ const useSyncer = () => {
   }
 }
 
-const NoteTreeItem: React.FC<{note: Note | NoteNode}> = ({note, children}) => {
+const NoteTreeItem: React.FC<{ note: Note | NoteNode }> = ({ note, children }) => {
   const [menuVisible, setMenuVisible] = useState(false);
   const [currentDeleteId, setCurrentDeleteId] = useState<string | null>(null);
+  const currentNoteId = useRecoilValue(currentNoteIdState);
   const noteNodesMap = useRecoilValue(noteNodesMapState);
+  const setNeedSave = useSetRecoilState(needSaveState);
   const [loading, withLoading] = useLoading()
   const setNotesState = useSetRecoilState(noteNodesMapState)
+  const checkNeedSave = useCheckNeedSave();
   const syncer = useSyncer();
   const handleAdd = async (currentId: NoteNode['id']) => {
+    if (checkNeedSave()) return;
     syncer.start(currentId)
-    const res = await withLoading(createDoc(id => createNoteNode(id, {pid: currentId})))
+    const res = await withLoading(createDoc(id => createNoteNode(id, { pid: currentId })))
     syncer.stop(currentId)
     setNotesState((origin) => produce(origin, draft => {
       const currentNoteNode = origin.get(currentId)!;
       draft.set(res.id, res);
-      draft.set(currentId, {...currentNoteNode, cids: [...currentNoteNode.cids, res.id]})
+      draft.set(currentId, { ...currentNoteNode, cids: [...currentNoteNode.cids, res.id] })
     }))
   }
   const handleDelete = async (currentId: NoteNode['id']) => {
+    const isCurrentDeleteIdAlsoCurrentNoteId = currentNoteId === currentId;
+    if (isCurrentDeleteIdAlsoCurrentNoteId && checkNeedSave()) return;
     const ids = syncer.start(currentId)
     await withLoading(deleteDocs(ids))
     syncer.stop(currentId)
@@ -208,13 +245,16 @@ const NoteTreeItem: React.FC<{note: Note | NoteNode}> = ({note, children}) => {
       noteNodesMapTraverse(noteNodesMap, currentId, id => draft.delete(id))
       if (currentNoteNode.pid) {
         const parentNoteNode = origin.get(currentNoteNode.pid)!;
-        draft.set(parentNoteNode.id, {...parentNoteNode, cids: parentNoteNode.cids.filter(cid => cid !== currentId)})
+        draft.set(parentNoteNode.id, { ...parentNoteNode, cids: parentNoteNode.cids.filter(cid => cid !== currentId) })
       }
     }))
+    if (isCurrentDeleteIdAlsoCurrentNoteId) {
+      setNeedSave(false);
+    }
   }
 
   const syncing = syncer.getSyncing(note.id);
-  console.log(syncing);
+
   return (
     <TreeItem
       nodeId={note.id}
@@ -227,8 +267,8 @@ const NoteTreeItem: React.FC<{note: Note | NoteNode}> = ({note, children}) => {
                 open={menuVisible}
                 onClose={() => setMenuVisible(false)}
               >
-                <MenuItem className="text-xs" onClick={() => {handleAdd(note.id); setMenuVisible(false)}}><Add className="mr-2 text-base"/>Add</MenuItem>
-                <MenuItem className="text-xs" onClick={(e) => {e.stopPropagation(); setCurrentDeleteId(note.id); setMenuVisible(false)}}><Delete className="mr-2 text-base"/>Delete</MenuItem>
+                <MenuItem className="text-xs" onClick={() => { handleAdd(note.id); setMenuVisible(false) }}><Add className="mr-2 text-base" />Add</MenuItem>
+                <MenuItem className="text-xs" onClick={(e) => { e.stopPropagation(); setCurrentDeleteId(note.id); setMenuVisible(false) }}><Delete className="mr-2 text-base" />Delete</MenuItem>
               </Menu>
             )}
           >
@@ -237,7 +277,7 @@ const NoteTreeItem: React.FC<{note: Note | NoteNode}> = ({note, children}) => {
               disabled={syncing}
               className={`p-0 min-w-0 transition-opacity ${(syncing || menuVisible) ? 'opacity-100' : 'opacity-0'} group-hover:opacity-100`}
               variant="text"
-              onClick={(e) => {e.stopPropagation(); setMenuVisible(true)}}
+              onClick={(e) => { e.stopPropagation(); setMenuVisible(true) }}
             >
               <MoreHorizIcon></MoreHorizIcon>
             </LoadingButton>
@@ -245,8 +285,8 @@ const NoteTreeItem: React.FC<{note: Note | NoteNode}> = ({note, children}) => {
           <Dialog open={currentDeleteId !== null} onClose={() => setCurrentDeleteId(null)}>
             <DialogTitle>Are you sure to delete the note and its children?</DialogTitle>
             <DialogActions>
-              <Button onClick={(e) => {e.stopPropagation(); setCurrentDeleteId(null)}}>Cancel</Button>
-              <Button color="error" onClick={(e) => {e.stopPropagation(); handleDelete(currentDeleteId!); setCurrentDeleteId(null);}}>Delete</Button>
+              <Button onClick={(e) => { e.stopPropagation(); setCurrentDeleteId(null) }}>Cancel</Button>
+              <Button color="error" onClick={(e) => { e.stopPropagation(); handleDelete(currentDeleteId!); setCurrentDeleteId(null); }}>Delete</Button>
             </DialogActions>
           </Dialog>
         </div>
@@ -261,45 +301,48 @@ const NoteTreeItem: React.FC<{note: Note | NoteNode}> = ({note, children}) => {
 const NotesTree: React.FC = () => {
   const noteNodesMap = useRecoilValue(noteNodesMapState)
   const notesTree = useRecoilValue(notesTreeSelector);
-  const setCurrentNoteId = useSetRecoilState(currentNoteIdState);
+  const [currentNoteId, setCurrentNoteId] = useRecoilState(currentNoteIdState);
   const searchValue = useRecoilValue(searchState);
+  const checkNeedSave = useCheckNeedSave();
 
   const renderTree = useCallback((notes: Array<Note>) => {
     if (notes.length === 0) return null;
     return notes.map(note => (
-        <NoteTreeItem
-          key={'tree' + note.id}
-          note={note}
-        >
-          {renderTree(note.children)}
-        </NoteTreeItem>
-      )
+      <NoteTreeItem
+        key={'tree' + note.id}
+        note={note}
+      >
+        {renderTree(note.children)}
+      </NoteTreeItem>
+    )
     );
   }, [])
 
   const renderMap = useCallback((noteNodes: NoteNodesMap) => {
     return [...noteNodes.values()].filter(noteNode => noteNode.name.includes(searchValue)).map(noteNode => (
-        <NoteTreeItem
-          key={'map' + noteNode.id}
-          note={noteNode}
-        >
-        </NoteTreeItem>
-      )
+      <NoteTreeItem
+        key={'map' + noteNode.id}
+        note={noteNode}
+      >
+      </NoteTreeItem>
+    )
     )
   }, [searchValue])
 
   const onTreeNodeSelect = (e: any, id: string) => {
+    if (checkNeedSave()) return;
     setCurrentNoteId(id);
   };
 
   return <div>
     <TreeView
+      selected={currentNoteId || undefined}
       defaultCollapseIcon={<ExpandMore />}
       defaultExpandIcon={<ChevronRight />}
       onNodeSelect={onTreeNodeSelect}
     >
       {
-        searchValue ? renderMap(noteNodesMap): renderTree(notesTree)
+        searchValue ? renderMap(noteNodesMap) : renderTree(notesTree)
       }
     </TreeView>
   </div>
@@ -308,22 +351,36 @@ const NotesTree: React.FC = () => {
 const Sider: React.FC = () => {
   return (
     <div className="h-screen w-72 bg-stone-800">
-      <ActionBar/>
-      <NotesTree/>
+      <ActionBar />
+      <NotesTree />
     </div>
   )
+}
+
+const useSave = () => {
+  const currentNote = useRecoilValue(currentNoteSelector);
+  const setNeedSave = useSetRecoilState(needSaveState);
+  const syncer = useSyncer();
+  return async () => {
+    if (!currentNote) return;
+    syncer.start(currentNote.id);
+    await saveDoc(currentNote.id, currentNote)
+    syncer.stop(currentNote.id);
+    setNeedSave(false);
+  }
 }
 
 const Save: React.FC = () => {
   const [successVisible, setSuccessVisible] = useState(false);
   const currentNote = useRecoilValue(currentNoteSelector);
   const syncer = useSyncer();
+  const { stop } = useAutoSave();
+  const save = useSave();
   const handleSave = async () => {
     if (!currentNote) return;
-    syncer.start(currentNote.id);
-    await saveDoc(currentNote.id, currentNote)
-    syncer.stop(currentNote.id);
+    await save();
     setSuccessVisible(true);
+    stop();
   }
   const loading = currentNote ? syncer.getSyncing(currentNote.id) : false;
   return (
@@ -348,16 +405,51 @@ const Save: React.FC = () => {
   )
 }
 
+const useAutoSave = () => {
+  const [countDown, setCountDown] = useRecoilState(autoSaveCountDownState);
+  const [autoSaveTimer, setAutoSaveTimer] = useRecoilState(autoSaveTimerState);
+
+  const clearAutoSaveTimer = () => {
+    if (autoSaveTimer) {
+      window.clearInterval(autoSaveTimer)
+      setAutoSaveTimer(null)
+    }
+  }
+
+  useEffect(() => {
+    if (countDown === 0 && autoSaveTimer) {
+      clearAutoSaveTimer()
+    }
+  }, [countDown])
+
+  return {
+    start: () => {
+      setCountDown(AUTO_SAVE_TIME)
+      clearAutoSaveTimer()
+      setAutoSaveTimer(window.setInterval(() => {
+        setCountDown(v => v - 1)
+      }, 1000))
+    },
+    stop: () => {
+      clearAutoSaveTimer();
+      setCountDown(AUTO_SAVE_TIME)
+    },
+    countDown,
+  }
+}
+
 const Content: React.FC = () => {
   const currentNoteId = useRecoilValue(currentNoteIdState);
   const currentNote = useRecoilValue(currentNoteSelector);
   const setNoteNodesMap = useSetRecoilState(noteNodesMapState);
   const codeEditorRef = useRef();
+  const setNeedSave = useSetRecoilState(needSaveState);
+  const { start } = useAutoSave();
 
   return (
     <div className="flex flex-1 bg-stone-700">
       {
-        currentNoteId ? <>
+        currentNote && currentNoteId ? <>
           <div className="relative flex-1 bg-stone-600">
             <CodeEditor
               ref={codeEditorRef}
@@ -368,13 +460,16 @@ const Content: React.FC = () => {
                 // @ts-ignore
                 const editorValue = codeEditorRef.current?.getValue?.()
                 setNoteNodesMap((origin) => produce(origin, draft => {
-                  draft.set(currentNoteId, {...draft.get(currentNoteId)!, name: getName(editorValue) || '', raw: editorValue})
+                  draft.set(currentNoteId, { ...draft.get(currentNoteId)!, name: getName(editorValue) || '', raw: editorValue })
                 }));
+                start();
+                setNeedSave(true)
               }}
             />
             <Save></Save>
+            <AutoSave></AutoSave>
           </div>
-          { currentNote ? <div className="flex-1 bg-stone-500"><MarkdownView value={currentNote.raw} /></div> : null }
+          {currentNote ? <div className="flex-1 bg-stone-500"><MarkdownView value={currentNote.raw} /></div> : null}
         </> : 'Pick a Note'
       }
     </div>
@@ -390,16 +485,46 @@ const useDataInitializer = () => {
   }, [])
 }
 
+const AutoSave: React.FC = () => {
+  const [open, setOpen] = useState(false);
+  const { countDown, stop } = useAutoSave();
+  const save = useSave();
+  useEffect(() => {
+    if (countDown <= 0) {
+      save()
+      setOpen(false)
+    } else if (countDown <= 10) {
+      setOpen(true)
+    } else {
+      setOpen(false)
+    }
+  }, [countDown])
+  return (
+    <Snackbar open={open} autoHideDuration={6000} transitionDuration={0}>
+      <Alert severity="info" sx={{ width: '100%' }} onClose={() => { setOpen(false); stop() }}>
+        auto save after {countDown}s
+      </Alert>
+    </Snackbar>
+  )
+}
+
 const Note: React.FC = () => {
   const snapshot = useRecoilSnapshot();
+  const needSave = useRecoilValue(needSaveState);
   useDataInitializer()
   useEffect(() => {
     console.debug('The following atoms were modified:');
-    for (const node of snapshot.getNodes_UNSTABLE({isModified: true})) {
+    for (const node of snapshot.getNodes_UNSTABLE({ isModified: true })) {
       console.debug(node.key, snapshot.getLoadable(node));
     }
   }, [snapshot]);
-  
+
+  useEffect(() => {
+    window.onbeforeunload = needSave ? function (e) {
+      e.returnValue = ("确定离开当前页面吗？");
+    } : null
+  }, [needSave]);
+
   return <div key="Note" className="flex min-h-screen text-slate-100 bg-stone-900">
     <Sider></Sider>
     <Content></Content>
@@ -412,7 +537,7 @@ const NotePage: NextPage = () => {
   return (
     isMe
       ? <RecoilRoot><Note /></RecoilRoot>
-      : <div key="password" className="flex min-h-screen justify-center items-center text-slate-100 bg-stone-900"><TextField label="password" variant="standard" onChange={e => setIsMe(e.target.value === 'taterbumb')}/></div>
+      : <div key="password" className="flex min-h-screen justify-center items-center text-slate-100 bg-stone-900"><TextField label="password" variant="standard" onChange={e => setIsMe(e.target.value === 'taterbumb')} /></div>
   )
 }
